Feature Implementation
Parsing the phrases/words based on speech has been implemented.

Arrays consist of phrases were made. Phrases to be replaced is are checked for such cases as position "a1" being recognized as "a one" or "a 1" to all be "a1" for the uniformity.
Recognized sentence are checked for trigger phrases regardless of the order of presence, so the phrases like "object a", "object b" and "object c", that require the presence order to identify which one is the target object and which one is the relative object, were tokenized to store the data of order.